+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [fuji] fix: should not display the `static` fields in `/fuji inspect configurations` command. (To hides the `not interested` fields.)
-- [core] fix: should remember the `parent GUI` when using the `search function` in `paged GUI`. (This makes the GUI interaction more convenient.)
+- [core] fix: should remember the `parent GUI` when using the `search function` in `paged GUI`. (This makes the GUI interaction more convenient, and the possibility for `recursive search`.)
 - [fuji] feature: now will use different items to represent `argument types` from `vanilla Minecraft` and `fuji`. (For `/fuji inspect argument-types` command)
+- [fuji] feature: now adds the `core` as a `dummy module` in `/fuji inspect modules` command, for convenience.